# go/stats: improve performance of safeJoinLabels