safeJoinLabels is in a pretty hot path for builtinbackupengine, every Read/Write file operation requires 2 calls to this function through (*scopedStats).TimedIncrementBytes.

An optimization pass to this function was pretty low hanging fruit compared to trying to materialize/cache the label values correctly.

```
$ benchstat {before,after}.txt
goos: darwin
goarch: arm64
pkg: vitess.io/vitess/go/stats
cpu: Apple M1 Max
                              │  before.txt  │              after.txt              │
                              │    sec/op    │   sec/op     vs base                │
SafeJoinLabels/no_combined-10   204.40n ± 1%   84.58n ± 0%  -58.62% (p=0.000 n=10)
SafeJoinLabels/combined-10       87.54n ± 9%   30.98n ± 1%  -64.60% (p=0.000 n=10)
geomean                          133.8n        51.19n       -61.73%

                              │ before.txt  │             after.txt              │
                              │    B/op     │    B/op     vs base                │
SafeJoinLabels/no_combined-10   152.00 ± 0%   48.00 ± 0%  -68.42% (p=0.000 n=10)
SafeJoinLabels/combined-10      104.00 ± 0%   24.00 ± 0%  -76.92% (p=0.000 n=10)
geomean                          125.7        33.94       -73.00%

                              │ before.txt │             after.txt              │
                              │ allocs/op  │ allocs/op   vs base                │
SafeJoinLabels/no_combined-10   4.000 ± 0%   1.000 ± 0%  -75.00% (p=0.000 n=10)
SafeJoinLabels/combined-10      2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
geomean                         2.828        1.000       -64.64%
```

This showed up as the top offender of allocations in pprofs I was looking at:
![image](https://github.com/user-attachments/assets/ece36885-106d-48b6-a080-9afdb04ed300)


## Related Issue(s)


## Checklist

-   [ ] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [ ] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required

## Deployment Notes

None